python_omr_checker/requirements.txt
# Python OMR Checker Requirements
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scikit-image==0.21.0
scipy==1.11.3
matplotlib==3.7.2
imutils==0.5.4
argparse==1.4.0
json5==0.9.14
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0